Reverted OkHttp dependency to 1.6.0, but now using an API that can be forced to use 2.0.0 immediately.

+++ extensions/robospice-okhttp-parent/robospice-okhttp/src/main/java/com/octo/android/robospice/request/okhttp/simple/OkHttpBigBinaryRequest.java
@@ -37,7 +37,7 @@ public class OkHttpBigBinaryRequest extends OkHttpBinaryRequest {
     }
 
     @Override
-    public InputStream processStream(final long contentLength,
+    public InputStream processStream(final int contentLength,
         final InputStream inputStream) throws IOException {
         OutputStream fileOutputStream = null;
         try {
